python/requirements.txt
PyHive
flask
google-cloud-dialogflow
thrift
thrift-sasl
google-api-core
sasl-0.3.1-cp310-cp310-win_amd64.whl